experimental/canvas/graphql/pyproject.toml
[tool.poetry]
name = "canvas-graphql-exploration"
version = "0.1.0"
description = ""
authors = ["Ed-Fi Alliance, LLC and contributors"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
python-dotenv = "^0.19.2"
pandas = "^1.4.1"
canvasapi = "^2.2.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
